requirements.txt
dotenv
google-genai == 1.50.1
opensearch-py
gradio